Move `log_msg` protos helper to the correct file
